IHDP-2 Tests refactor
