+++ core/src/test/java/org/infinispan/hadoop/testutils/mapreduce/HDFSReducer.java
@@ -1,0 +1,23 @@
+package org.infinispan.hadoop.testutils.mapreduce;
+
+import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.Text;
+import org.apache.hadoop.mapreduce.Reducer;
+
+import java.io.IOException;
+
+/**
+ * Reducer using Hadoop writables only
+ */
+public class HDFSReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+   private IntWritable result = new IntWritable();
+
+   public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+      int sum = 0;
+      for (IntWritable val : values) {
+         sum += val.get();
+      }
+      result.set(sum);
+      context.write(key, result);
+   }
+}
